[KOGITO-610] Adding persistence=required to image labels
