+++ kogito-codegen/src/main/java/org/kie/kogito/codegen/process/persistence/PersistenceGenerator.java
@@ -53,6 +53,7 @@ import org.kie.kogito.codegen.GeneratedFile;
 import org.kie.kogito.codegen.di.DependencyInjectionAnnotator;
 import org.kie.kogito.codegen.metadata.MetaDataWriter;
 import org.kie.kogito.codegen.metadata.PersistenceLabeler;
+import org.kie.kogito.codegen.metadata.PersistenceProtoFilesLabeler;
 import org.kie.kogito.codegen.process.persistence.proto.Proto;
 import org.kie.kogito.codegen.process.persistence.proto.ProtoGenerator;
 
@@ -74,6 +75,7 @@ public class PersistenceGenerator extends AbstractGenerator {
     
     private ClassLoader classLoader;
     
+    private PersistenceProtoFilesLabeler persistenceProtoLabeler = new PersistenceProtoFilesLabeler();
     private PersistenceLabeler persistenceLabeler = new PersistenceLabeler();
         
     public PersistenceGenerator(File targetDirectory, Collection<?> modelClasses, boolean persistence, ProtoGenerator<?> protoGenerator, List<String> parameters) {
@@ -87,7 +89,10 @@ public class PersistenceGenerator extends AbstractGenerator {
         this.protoGenerator = protoGenerator;
         this.classLoader = classLoader;
         this.parameters = parameters;
-        this.addLabeler(persistenceLabeler);
+        if (this.persistence) {
+            this.addLabeler(persistenceProtoLabeler);
+            this.addLabeler(persistenceLabeler);
+        }
     }
 
     @Override
@@ -107,8 +112,8 @@ public class PersistenceGenerator extends AbstractGenerator {
             if (persistencePath != null && persistencePath.isDirectory()) {
                 // only process proto files generated by the inner generator
                 for (final File protoFile : Objects.requireNonNull(persistencePath.listFiles((dir, name) ->
-                        !KOGITO_APPLICATION_PROTO.equalsIgnoreCase(name) && name.toLowerCase().endsWith(PersistenceLabeler.PROTO_FILE_EXT))))
-                    this.persistenceLabeler.processProto(protoFile);
+                        !KOGITO_APPLICATION_PROTO.equalsIgnoreCase(name) && name.toLowerCase().endsWith(PersistenceProtoFilesLabeler.PROTO_FILE_EXT))))
+                    this.persistenceProtoLabeler.processProto(protoFile);
             }
             
             if (!protoFilePath.toFile().exists()) {
